covid_api/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-countries==7.2.1
djangorestframework==3.12.4
idna==3.2
kaleido==0.2.1
plotly==5.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tenacity==8.0.1
urllib3==1.26.6
